[feat] 판매자 앱의 서비스 상세 페이지 navigation 구현

+++ core/ui/src/main/kotlin/com/ssavice/ui/SellerServiceListItem.kt
@@ -41,7 +41,7 @@ fun SellerServiceListItem(
     modifier: Modifier = Modifier,
     cornerRadius: Dp = 14.dp,
     elevation: Dp = 8.dp,
-    onClick: ((Int) -> Unit)? = null,
+    onClick: ((Long) -> Unit)? = null,
     thumbnail: @Composable () -> Unit,
 ) {
     val shape = RoundedCornerShape(cornerRadius)
